backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
requests
google-generativeai
python-jose[cryptography]
dramatiq[redis]